+++ Src/Library/changelog.md
@@ -7,3 +7,4 @@
 ### FIXES
 - double response issue if pre-processor sent response and validator also has failures. [#info](https://discord.com/channels/933662816458645504/1080609437879914506)
 - `EndpointWithMapping<TRequest, TResponse, TEntity>` map methods becoming protected in v5.8 [#info](https://discord.com/channels/933662816458645504/1082207914376319026)
+- endpoint factory being inferred from request body issue [#info](https://discord.com/channels/933662816458645504/1084841217898061915)